Add and configure emacs lsp-mode
This works by having nodePackages.typescript-language-server installed
through the shell.nix file in the demando project. It uses the official
tsserver under the hood, which it grabs from the local node_modules
directory.

This also replaces the auto-complete package with company, as lsp-mode
automatically picks it up for its completions